templates/ecommerce-assistant/pyproject.toml
[project]
name = "ecommerce-assistant-agent"
version = "1.0.0"
description = "AI-powered e-commerce assistant agent template using CrewAI. Searches products, tracks orders, handles returns, and recommends items."
readme = "README.md"
requires-python = ">=3.10,<3.14"
license = { text = "MIT" }
authors = [{ name = "AI Agent Templates" }]

dependencies = [
    "crewai>=1.9,<2.0",
    "crewai-tools>=0.36",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
